+++ client_test.go
@@ -102,7 +102,7 @@ func TestClientNamespaces(t *testing.T) {
 }
 
 func TestVersion(t *testing.T) {
-	if VERSION != "2.0.0" {
-		t.Errorf("VERSION = %s, want 2.0.0", VERSION)
+	if VERSION != "2.0.1" {
+		t.Errorf("VERSION = %s, want 2.0.1", VERSION)
 	}
 }